Caffe references throughout code
The comments and the code refer to Caffe even TensorFlow is the language we're focusing on.

I haven't changed the file type for "protobuf"; however TensorFlow outputs a .ckpt file not .pb, does the file need to be converted or is this a mistake? Have not made all changes as I'm unsure what the library requires.